backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Supabase
supabase==2.0.2
postgrest==0.13.0

# Redis
redis==5.0.1
celery==5.3.4

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI and ML
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2
anthropic==0.7.7

# File Processing
python-magic==0.4.27
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Audio Processing
whisper==1.1.10
librosa==0.10.1

# Image Processing
opencv-python==4.8.1.78
pytesseract==0.3.10

# BPMN Processing
bpmn-python==0.0.1

# Notifications
smtplib
requests==2.31.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Azure Integration
azure-identity==1.15.0
azure-graphrbac==0.61.1

# External Tools
n8n-python==0.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3 